Move user management resource types
