+++ tests/utils/fixtures/mock_organization_membership.py
@@ -1,6 +1,6 @@
 import datetime
 
-from workos.resources.user_management import OrganizationMembership
+from workos.types.user_management import OrganizationMembership
 
 
 class MockOrganizationMembership(OrganizationMembership):
