+++ workos/types/events/event.py
@@ -1,7 +1,7 @@
 from typing import Literal, Union
 from pydantic import Field
 from typing_extensions import Annotated
-from workos.resources.user_management import OrganizationMembership, User
+from workos.types.user_management import OrganizationMembership, User
 from workos.types.directory_sync.directory_group import DirectoryGroup
 from workos.types.directory_sync.directory_user import DirectoryUser
 from workos.types.events.authentication_payload import (
@@ -39,12 +39,12 @@ from workos.types.organizations.organization_common import OrganizationCommon
 from workos.types.organizations.organization_domain import OrganizationDomain
 from workos.types.roles.role import Role
 from workos.types.sso.connection import Connection
-from workos.types.user_management.email_verification_common import (
+from workos.types.user_management.email_verification import (
     EmailVerificationCommon,
 )
-from workos.types.user_management.invitation_common import InvitationCommon
-from workos.types.user_management.magic_auth_common import MagicAuthCommon
-from workos.types.user_management.password_reset_common import PasswordResetCommon
+from workos.types.user_management.invitation import InvitationCommon
+from workos.types.user_management.magic_auth import MagicAuthCommon
+from workos.types.user_management.password_reset import PasswordResetCommon
 
 
 class AuthenticationEmailVerificationSucceededEvent(
