+++ workos/types/user_management/password_reset.py
@@ -9,3 +9,10 @@ class PasswordResetCommon(WorkOSModel):
     email: str
     expires_at: str
     created_at: str
+
+
+class PasswordReset(PasswordResetCommon):
+    """Representation of a WorkOS PasswordReset object."""
+
+    password_reset_token: str
+    password_reset_url: str
