+++ workos/types/webhooks/webhook.py
@@ -3,7 +3,7 @@ from pydantic import Field
 from typing_extensions import Annotated
 from workos.types.directory_sync import DirectoryGroup
 from workos.types.events import EventPayload
-from workos.resources.user_management import OrganizationMembership, User
+from workos.types.user_management import OrganizationMembership, User
 from workos.types.webhooks.webhook_model import WebhookModel
 from workos.types.workos_model import WorkOSModel
 from workos.types.directory_sync.directory_user import DirectoryUser
@@ -41,12 +41,12 @@ from workos.types.organizations.organization_common import OrganizationCommon
 from workos.types.organizations.organization_domain import OrganizationDomain
 from workos.types.roles.role import Role
 from workos.types.sso.connection import Connection
-from workos.types.user_management.email_verification_common import (
+from workos.types.user_management.email_verification import (
     EmailVerificationCommon,
 )
-from workos.types.user_management.invitation_common import InvitationCommon
-from workos.types.user_management.magic_auth_common import MagicAuthCommon
-from workos.types.user_management.password_reset_common import PasswordResetCommon
+from workos.types.user_management.invitation import InvitationCommon
+from workos.types.user_management.magic_auth import MagicAuthCommon
+from workos.types.user_management.password_reset import PasswordResetCommon
 
 
 class AuthenticationEmailVerificationSucceededWebhook(
